ai-qa-agent/requirements.txt
streamlit>=1.28.0
appium-python-client>=3.1.0
selenium>=4.15.0
openai>=1.6.0
pydantic>=2.5.0
python-dotenv>=1.0.0
pillow>=10.0.0
